backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database (Async PostgreSQL with SQLAlchemy 2.0)
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Authentication
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0

# Environment & Configuration
python-dotenv>=1.0.1

# Pydantic (already included with FastAPI, but explicitly pinning)
pydantic>=2.10.0
pydantic-settings>=2.6.0

